requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio-nats-client==0.11.5
attrs==21.2.0
charset-normalizer==2.0.9
frozenlist==1.2.0
idna==3.3
j2tools==2019.3
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
motor==2.5.1
multidict==5.2.0
orjson==3.6.5
Pillow==8.4.0
pydantic==1.8.2
pymongo==3.12.3
pyzbar==0.1.8
qrcode==7.3.1
rocketgram
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
typing_extensions==4.0.1
umongo==3.0.1
yarl==1.7.2
